hmi-controller: Added IVI_HMI_CONTROLLER_LAYOUT_MODE_ORIG
While we have 4 existing policies for layout modes all of them
seem to be resizing the clients buffers to some layer based values.
Sometimes it may not look good to resize a client's buffers and
may make more sense to leave the buffers untouched. With this
additional policy, we now have a way to present the applications
buffers as the application wanted without any reconfiguration,
starting off at 0,0.

There can always be run time changes to the configuration if
need be from either the client or compositor side.

Signed-off-by: Satyeshwar Singh <satyeshwar.singh@intel.com>
Signed-off-by: Nobuhiko Tanibata <NOBUHIKO_TANIBATA@xddp.denso.co.jp>